+++ ICD.Connect.Audio.Biamp/BiampTesiraDevice.cs
@@ -29,9 +29,12 @@ using ICD.Connect.Protocol.Ports;
 using ICD.Connect.Protocol.Ports.ComPort;
 using ICD.Connect.Protocol.Settings;
 using ICD.Connect.Settings;
+using ICD.Connect.Settings.Core;
+using ICD.Connect.Telemetry.Attributes;
 
 namespace ICD.Connect.Audio.Biamp
 {
+	[ExternalTelemetry("Biamp", typeof(BiampExternalTelemetryProvider))]
 	public sealed class BiampTesiraDevice : AbstractDevice<BiampTesiraDeviceSettings>
 	{
 		// How often to re-subscribe to device events