+++ ICD.Connect.Audio/DspTelemetryNames.cs
@@ -1,0 +1,20 @@
+﻿namespace ICD.Connect.Audio
+{
+	public static class DspTelemetryNames
+	{
+		public const string CALL_ACTIVE = "Dsp Call Active";
+		public const string CALL_ACTIVE_CHANGED = "Dsp Call Active Changed";
+
+		public const string ACTIVE_FAULTS = "Dsp Active Faults";
+		public const string ACTIVE_FAULTS_CHANGED = "Dsp Active Faults Changed";
+
+		public const string FIRMWARE_VERSION = "Dsp Firmware Version";
+		public const string FIRMWARE_VERSION_CHANGED = "Dsp Firmware Version Changed";
+
+		public const string VOIP_REGISTERED = "Dsp Voip Registered";
+		public const string VOIP_REGISTERED_CHANGED = "Dsp Voip Registered Changed";
+
+		public const string VOIP_REGISTRATION_STATUS = "Dsp Voip Registration Status";
+		public const string VOIP_REGISTRATION_STATUS_CHANGED = "Dsp Voip Registration Status Changed";
+	}
+}